requirements.txt
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==10.4.0
numpy==1.26.4
streamlit==1.39.0
pandas==2.2.3
python-dotenv==1.0.1
